# [internal] Improve Go performance by running fewer processes to analyze external packages