--- a/src/python/pants/backend/go/util_rules/external_pkg.py
+++ b/src/python/pants/backend/go/util_rules/external_pkg.py
@@ -1,6 +1,8 @@
 # Copyright 2021 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
+from __future__ import annotations
+
 import os
 from dataclasses import dataclass
 from typing import Tuple