… src/python/pants/backend/experimental/go/register.py
… src/python/pants/backend/go/goals/package_binary_integration_test.py
… src/python/pants/backend/go/util_rules/assembly_integration_test.py
… src/python/pants/backend/go/util_rules/build_go_pkg_test.py
… src/python/pants/backend/go/util_rules/external_pkg_test.py