@tdyas 2021-10-06
> Instead, we can simply run it once per module and leverage the engine's memoization to avoid re-running.

Does this rely on `pantsd` running?