@Eric-Arellano 2021-10-06
> Does this rely on pantsd running?

No, even if you don't use pantsd the engine will memoize for the current run. See the benchmarks I just posted without Pantsd :D