@tdyas 2021-10-06
> No, even if you don't use pantsd the engine will memoize for the current run. See the benchmarks I just posted without Pantsd :D

What about subsequent runs? Would this get cached remotely in a CAS?