@Eric-Arellano 2021-10-06
> What about subsequent runs? Would this get cached remotely in a CAS?

The `go list -json` `Process` gets cached with local cache and remote cache. The parsing of the JSON result only gets memoized, but it's also fast.

This is the same with target generation. The underlying `Process` required is cached to LMDB, and then the creation of the targets in-memory is memoized.